Code Review :)
Hey Anthony, TA Alan here. So, first off I'd like to compliment you on your use of jinja templating, great use of some of the features not covered on the platform in depth. Additionally, your server.py is fairly clean and easy to read which is nice. But I have one idea that could make your wall.html like 1/3 the size. Currently you are importing 'wall_content' and doing an admirable manipulating the bajeezus out of it. 2 things 1) In my opinion, it makes your front-end hard to read and you repeat yourself(which is bad) and 2)your query that you make to create wall_content uses 2 joins, which are very expensive operations to make in SQL. My approach to this was to make two seperate queries to the db, one for comments and one for users. This allows clean logic that Ioutline on lines 62-70 in wall.html. DM on slack if you have more questions